pyproject.toml
[tool.poetry]
name = "auth-onedrive"
version = "0.1.0"
description = ""
authors = ["Christian Bueno <chmabuen@espol.edu.ec>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
msal = "^1.30.0"
requests = "^2.32.3"
msgraph-sdk = "^1.5.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "auth_onedrive.main:run"
start2 = "auth_onedrive.main2:run"
